requirements.txt
transformers>=4.36.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
peft>=0.6.2
gradio>=3.47.1
fastapi>=0.105.0
uvicorn>=0.25.0
requests>=2.31.0
datasets>=2.16.0
pytest>=7.4.0
torch>=2.1.0
numpy>=1.26.0
httpx>=0.24
